Update README.md file
I had to update my README file in GitHub because when I pushed the changes I made in my local environment, the image did not show up. I could see it in the preview in VS code, but I couldn't see it on here, so I edited this file in GitHub to fix this issue.